Run `composer bump`
Given that PIE is a project rather than a library and is distributed as a Phar
archive, we do not need to keep broad compatibility with older library
versions. Run `composer bump` to prevent accidental downgrades of libraries, by
increasing the versions in `composer.json` to match the currently locked
versions in `composer.lock`.