Configure fine-grained GitHub Action permissions
This change configures GitHub Actions to only allow reading the repository
contents within the top-level workflow-wide permissions section. Extended
permissions are then granted to individual jobs to make permissions “secure by
default” even when additional jobs are added to an existing workflow.

As a follow-up to this PR, the “Workflow Permissions” in the repository
settings should be reconfigured to “Read repository contents and packages
permissions” instead of “Read and write permissions” to also make additional
workflows secure by default, even when a permissions section is not explicitly
configured.